Applied Java conventions to IO package (Reader/Writer, Input/OutputStream for ASCII, binary IO). Reduced debug output for non-SPD warning with full covariances. Fixed IOUtil.writeFloat for arrays.

+++ trunk/jstk/src/de/fau/cs/jstk/app/Map.java
@@ -183,7 +183,7 @@ public class Map {
 				List<Sample> data = new LinkedList<Sample>();
 				double [] buf = new double [fr.getFrameSize()];
 				while (fr.read(buf))
-					data.add(new Sample(0, buf));
+					data.add(new Sample((short) 0, buf));
 				fr.close();
 				
 				logger.info("Map.main(): " + data.size() + " samples read");
